requirements.txt
fastapi
pydantic_core
uvicorn
sentry_sdk
celery
pytest
pyamqp
requests
SQLAlchemy
python-dotenv
uuid-utils

pandas
psycopg[binary]
